requirements.txt
numpy==2.1.1
pandas==2.2.3
PyMySQL==1.1.1
streamlit==1.39.0
